fashion_store/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cloudinary==1.41.0
Django==5.1
django-ckeditor==6.7.1
django-js-asset==2.2.0
django-seed==0.3.1
djangorestframework==3.15.2
drf-yasg==1.21.7
Faker==28.0.0
graphviz==0.20.3
inflection==0.5.1
oauth2_provider==0.0
packaging==24.1
pillow==10.4.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
setuptools==73.0.1
six==1.16.0
sqlparse==0.5.1
toposort==1.10
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
